added userId to contacts

+++ api/contacts/data.js
@@ -14,6 +14,7 @@ class ContactsData {
 
   ///////////////////// "/"
   getMany(query) {
+    query.userId = req.user.id;
     return db.find(query)
       .sort({name: 1})
       .toArray();
@@ -22,6 +23,7 @@ class ContactsData {
   addOne(contact) {
     // add properties we expect to be there like mongoose would
     contact.id = contact.id || chance.guid();
+    contact.userId = req.user.id;
     contact.labels = contact.labels || [];
     return db.insertOne(contact)
       .then(result => {
@@ -31,26 +33,27 @@ class ContactsData {
 
   updateMany(contacts) {
     const arr = [];
-    contacts.forEach(contact => arr.push(db.updateOne({id: contact.id}, contact)))
+    contacts.forEach(contact => arr.push(db.updateOne({userId: req.user.id, id: contact.id}, contact)))
     return Promise.all(arr);
   }
 
   deleteMany(query) {
+    query.userId = req.user.id;
     return db.removeMany(query);
   }
 
   ///////////////////// "/:id"
   getOne(id) {
-    return db.findOne({id: id})
+    return db.findOne({userId: req.user.id, id: id})
   }
 
   updateOne(id, contact) {
     contact._id = ObjectId(contact._id);
-    return db.updateOne({id: id}, contact)
+    return db.updateOne({userId: req.user.id, id: id}, contact)
   }
 
   deleteOne(id) {
-    return db.removeOne({id: id});
+    return db.removeOne({userId: req.user.id, id: id});
   }
 
 }
